+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+- Integrated `pdftotext`, `pdfminer` and `docx2txt` interfaces into `filecontent` filter.
+- Removed `textract` and ~50 MB of dependencies as they are no longer needed.
+- Python 3.12 support
+
 ## v3.1.2 (2024-02-16)
 
 - Fixes a validation error where correctly defined actions were not accepted in Python 3.12.2.